moves everything from newer.rs into parent
